refactor(store): move auth module into its own folder
