+++ src/components/Login.vue
@@ -2,9 +2,15 @@
   <div id="login">
     <h2>Connexion</h2>
     <form @submit.prevent="login">
-      <div v-if="!isEmailValid" class="error" id="email-error">Please enter a valid email.</div>
-      <div v-if="!isPasswordValid" class="error" id="password-error">Please enter a password.</div>
-      <div v-if="loginSuccess" class="success">You've logged in successfully !!</div>
+      <div v-if="!isEmailValid" class="error" id="email-error">
+        Please enter a valid email.
+      </div>
+      <div v-if="!isPasswordValid" class="error" id="password-error">
+        Please enter a password.
+      </div>
+      <div v-if="loginSuccess" class="success">
+        You've logged in successfully !!
+      </div>
       <input type="text" placeholder="Email" v-model="email" />
       <input type="password" placeholder="Password" v-model="password" />
       <input type="submit" value="Se connecter" />
@@ -19,7 +25,7 @@ import { getModule } from 'vuex-module-decorators';
 
 import { login } from '@/api/user';
 import { validateEmail } from '@/utils/form-validation';
-import { AuthModule } from '@/store/AuthModule';
+import { AuthModule } from '@/store/auth/AuthModule';
 
 @Component
 export default class Login extends Vue {
@@ -36,7 +42,7 @@ export default class Login extends Vue {
 
     const { email, password } = this;
 
-    this.authModule.loginAction({email, password}).then(() => {
+    this.authModule.loginAction({ email, password }).then(() => {
       this.$router.replace('/lists');
     });
   }
