+++ src/store/auth/payloads.ts
@@ -1,0 +1,4 @@
+export interface AuthPayload {
+  email: string
+  password: string
+}
