+++ src/views/Home.vue
@@ -15,7 +15,7 @@ import Signup from '@/components/Signup.vue';
 import Login from '@/components/Login.vue';
 import CreateList from '@/components/CreateList.vue';
 import { ListModule } from '@/store/list/ListModule';
-import { AuthModule } from '@/store/AuthModule';
+import { AuthModule } from '@/store/auth/AuthModule';
 
 @Component({
   components: {
